+++ packages/uniforms/__tests__/index.ts
@@ -1,5 +1,4 @@
 import * as uniforms from 'uniforms';
-import * as suites from 'uniforms/__suites__';
 
 it('exports everything', () => {
   expect(uniforms).toMatchObject({
@@ -18,8 +17,4 @@ it('exports everything', () => {
     useField: expect.any(Function),
     useForm: expect.any(Function),
   });
-});
-
-describe('@RTL', () => {
-  suites.testConnectField();
 });
